refactor(github): rewrite CommitsTab and IssuesTab components
- Extract common UI components into separate files
- Use AnalysisTabLayout as the main container for analysis tabs
- Replace Accordion with UserTabItem for individual user data
- Use TabDataTable for displaying data tables
- Remove redundant code and simplify components
- Add themes for consistent styling
